spilt up files
